refactor(framework): one home for the skill contract
The contract lived in three disagreeing places: skill-authoring.md, the two
templates, and aidd_docs/CONTRIBUTING.md. A skill could satisfy one and break
another, so "divergence" had no truth value.

- R1 names the action table and forbids a rule in the Actions lead-in.
- R4 makes argument-hint name what the user brings, never the action slugs.
- R10 says what it meant: the router carries nothing an action could carry.
  Its old wording, "the leanest file", produced five false positives.
- R12 makes ## Test a Case/Pass table.

skill-generate broke three of its own rules and now follows them. CONTRIBUTING
routes every artifact to its generator instead of restating the contract.

sync-skill-argument-hints.mjs wrote argument-hint from action file names, kept a
hardcoded exception list, and staged the result on every commit. It was the
contract's third home, in code. check-skill-argument-hints.mjs replaces it and
refuses a hint that repeats the action slugs.
