+++ plugins/aidd-context/skills/04-skill-generate/assets/action-template.md
@@ -17,4 +17,6 @@
 
 ## Test
 
-<Deterministic checks, one per line.>
+| Case | Pass |
+| --- | --- |
+| <what is run> | <the observable result it gives> |
